feat(feeds): rename feeds page
