+++ realtime_feed_ui/app/containers/Feeds/reducer.js
